Minor capitalization and pluralization fixes
Actually, all "BeagleBoard" boards are ARM, but not all Beagles. We need
to find better encapsulating terms. Specifically, this page only
addresses "BeagleBoard" boards, like BeagleBoard, BeagleBoard-xM, and
BeagleBoard-X15, not BeaglePlay, BeagleBone, BeagleConnect or BeagleV boards.

We've started using the term "Beagles" to refer to the entirety of the
product line.

BeagleBoard is meant to be embedded single board computers that are
low-power and small-size. Both xM and X15 added more upon this base.
They are all ARM+DSP based boards.

BeagleBone is meant to have the cape headers for embedded prototyping.
Originally, BeagleBone stood for "the bare-bones BeagleBoard" where less
was integrated onto the board and more header pins were available to
make it more optimized for embedded systems. We also dropped the DSP
from the original BeagleBoard. AI and AI-64 bring back the DSP for
performing AI tasks.

BeaglePlay is meant to be an easier-to-use computer with standard
connections and integrated connectivity.

BeagleConnect is meant to be MCUs that connect to Linux computers to act
as the gateway or system controller.

BeagleV is meant to be RISC-V based boards.

Let's figure out a place to put this information and get ourselves
aligned to it or correct it as needed.

cc @lorforlinux