Confusion about where to put add-on usage docs
The "expansion" section is meant to be about how to *build* expansion
hardware. This is where we provide information about what we expect
people to do when building add-on hardware, not about telling people how
to use it. They can infer how to use it, but that should really be
broken down into the "quick start", "design" and "demos" sections.