Confusion about how to link to documents
The ToC needs to have links to the *files* and that determines the
location of the article in the overall document structure.

Do not confuse this with internal reference *links* that are based on
the names. Although links get made if the ToC isn't hidden, that isn't
the way to create the references in the ToC.

Refer to https://openbeagle.org/docs/docs.beagleboard.io/-/merge_requests/105